Split log parser
