+++ src/main.ts
@@ -17,7 +17,8 @@ import {Counter} from './components/counter'
 import {TeXMagician} from './components/texmagician'
 import {EnvPair} from './components/envpair'
 import {Section} from './components/section'
-import {Parser as LogParser} from './components/parser/log'
+import {LatexLogParser} from './components/parser/latexlog'
+import {LinterLogParser} from './components/parser/linterlog'
 import {UtensilsParser as PEGParser} from './components/parser/syntax'
 
 import {Completer} from './providers/completion'
@@ -340,7 +341,8 @@ export class Extension {
     readonly viewer: Viewer
     readonly server: Server
     readonly locator: Locator
-    readonly logParser: LogParser
+    readonly latexLogParser: LatexLogParser
+    readonly linterLogParser: LinterLogParser
     readonly pegParser: PEGParser
     readonly completer: Completer
     readonly linter: Linter
@@ -373,7 +375,8 @@ export class Extension {
         this.viewer = new Viewer(this)
         this.server = new Server(this)
         this.locator = new Locator(this)
-        this.logParser = new LogParser(this)
+        this.latexLogParser = new LatexLogParser(this)
+        this.linterLogParser = new LinterLogParser(this)
         this.completer = new Completer(this)
         this.linter = new Linter(this)
         this.cleaner = new Cleaner(this)
